pyproject.toml
[tool.poetry]
name = "cryoet-protein-identification"
version = "0.1.0"
description = "just a sample project to learn mlops tools"
authors = ["veta r <fetveta@gmail.com>"]
readme = "README.md"
packages = [
    {include = "cryoet-protein-identification", from = "."}
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pandas = "^2.3.2"
pytorch-lightning = "^2.5.4"
zarr = "<3.1.2"
copick-utils = {git = "https://github.com/copick/copick-utils.git", rev = "tomo-writer"}
matplotlib = "^3.10.6"
tqdm = "^4.67.1"
copick = "^1.12.0"
monai-weekly = {extras = ["mlflow"], version = "^1.6.dev2535"}
pathlib = "^1.0.1"
kaggle = "^1.7.4.5"
hydra-core = "^1.3.2"
torchinfo = "^1.8.0"
connected-components-3d = "^3.24.0"
traitlets = "^5.14.3"
albumentations = "^2.0.8"
transformers = "^4.56.1"
wandb = "^0.21.3"
onnx = "^1.19.0"


[tool.poetry.group.dev.dependencies]
fire = "^0.7.1"
pre-commit = "^4.3.0"
dvc = "^3.63.0"
dvc-gdrive = "^3.0.1"
pyopenssl = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
